# ui/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AssignmentRow": {
        "description": "One role a doctor was assigned on a specific date.\n\n`role` follows the v1 export convention:\n  - STATION_<station>_<AM|PM>\n  - ONCALL\n  - WEEKEND_EXT\n  - WEEKEND_CONSULT",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "doctor": {
            "title": "Doctor",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "doctor",
          "date",
          "role"
        ],
        "title": "AssignmentRow",
        "type": "object"
      },
...
      "BatchRunRequest": {
        "description": "One batch = current session state \u00d7 listed solvers \u00d7 listed seeds.",
        "properties": {
          "run_config": {
            "$ref": "#/components/schemas/RunConfig"
          },
          "seeds": {
            "items": {
              "type": "integer"
            },
            "title": "Seeds",
            "type": "array"
          },
          "solvers": {
            "items": {
              "enum": [
                "cpsat",
                "greedy",
                "random_repair"
              ],
              "type": "string"
            },
            "title": "Solvers",
            "type": "array"
          }
        },
        "title": "BatchRunRequest",
        "type": "object"
      },
      "BatchSummary": {
        "description": "Aggregate of one batch run. Quality ratio vs the first baseline\nin the request is reported as `quality_ratio_vs_*` if both solvers\nran on at least one shared seed.",
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feasibility_rate": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Feasibility Rate",
            "type": "object"
          },
          "instance_label": {
            "title": "Instance Label",
            "type": "string"
          },
          "mean_objective": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Mean Objective",
            "type": "object"
          },
          "mean_shortfall": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Mean Shortfall",
            "type": "object"
          },
          "n_days": {
            "title": "N Days",
            "type": "integer"
          },
          "n_doctors": {
            "title": "N Doctors",
            "type": "integer"
          },
          "n_stations": {
            "title": "N Stations",
            "type": "integer"
          },
          "quality_ratios": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Quality Ratios",
            "type": "object"
          },
          "run_config": {
            "$ref": "#/components/schemas/RunConfig"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/SingleRun"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "batch_id",
          "created_at",
          "instance_label",
          "n_doctors",
          "n_stations",
          "n_days",
          "run_config",
          "runs"
        ],
        "title": "BatchSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunConfig": {
        "description": "CP-SAT knobs. Phase 2 exposes only `time_limit_s`, `num_workers`,\n`random_seed`, and `feasibility_only`; Phase 3 extends this with\nbranching / linearization / presolve levers.",
        "properties": {
          "feasibility_only": {
            "default": false,
            "title": "Feasibility Only",
            "type": "boolean"
          },
          "num_workers": {
            "default": 1,
            "title": "Num Workers",
            "type": "integer"
          },
          "random_seed": {
            "default": 0,
            "title": "Random Seed",
            "type": "integer"
          },
          "time_limit_s": {
            "default": 30.0,
            "title": "Time Limit S",
            "type": "number"
          }
        },
        "title": "RunConfig",
        "type": "object"
      },
...
      "SelfCheckViolation": {
        "properties": {
          "location": {
            "title": "Location",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "rule",
          "severity",
          "location",
          "message"
        ],
        "title": "SelfCheckViolation",
        "type": "object"
      },
...
      "SingleRun": {
        "description": "One cell of (solver, seed) in the batch's cross product.",
        "properties": {
          "best_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Bound"
          },
          "coverage_over": {
            "title": "Coverage Over",
            "type": "integer"
          },
          "coverage_shortfall": {
            "title": "Coverage Shortfall",
            "type": "integer"
          },
          "first_feasible_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Feasible S"
          },
          "headroom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headroom"
          },
          "n_assignments": {
            "title": "N Assignments",
            "type": "integer"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "objective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "self_check_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Check Ok"
          },
          "solver": {
            "enum": [
              "cpsat",
              "greedy",
              "random_repair"
            ],
            "title": "Solver",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "violation_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Violation Count"
          },
          "wall_time_s": {
            "title": "Wall Time S",
            "type": "number"
          }
        },
        "required": [
          "run_id",
          "solver",
          "seed",
          "status",
          "wall_time_s",
          "objective",
          "best_bound",
          "headroom",
          "first_feasible_s",
          "self_check_ok",
          "violation_count",
          "coverage_shortfall",
          "coverage_over",
          "n_assignments"
        ],
        "title": "SingleRun",
        "type": "object"
      },
      "SingleRunDetail": {
        "description": "Full payload for one run \u2014 the assignments + self-check are kept\nhere rather than on `SingleRun` so the benchmark-table response\nstays lightweight.",
        "properties": {
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "coverage": {
            "additionalProperties": true,
            "title": "Coverage",
            "type": "object"
          },
          "fairness": {
            "additionalProperties": true,
            "title": "Fairness",
            "type": "object"
          },
          "result": {
            "$ref": "#/components/schemas/SolveResultPayload"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "solver": {
            "enum": [
              "cpsat",
              "greedy",
              "random_repair"
            ],
            "title": "Solver",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "batch_id",
          "solver",
          "seed",
          "result"
        ],
        "title": "SingleRunDetail",
        "type": "object"
      },
...
      "SolveEvent": {
        "description": "One improving solution found by CP-SAT.",
        "properties": {
          "assignments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AssignmentRow"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignments"
          },
          "best_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Bound"
          },
          "components": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Components",
            "type": "object"
          },
          "objective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "type": {
            "const": "event",
            "default": "event",
            "title": "Type",
            "type": "string"
          },
          "wall_s": {
            "title": "Wall S",
            "type": "number"
          }
        },
        "required": [
          "wall_s"
        ],
        "title": "SolveEvent",
        "type": "object"
      },
      "SolveResultPayload": {
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/AssignmentRow"
            },
            "title": "Assignments",
            "type": "array"
          },
          "best_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Bound"
          },
          "first_feasible_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Feasible S"
          },
          "intermediate": {
            "items": {
              "$ref": "#/components/schemas/SolveEvent"
            },
            "title": "Intermediate",
            "type": "array"
          },
          "n_constraints": {
            "title": "N Constraints",
            "type": "integer"
          },
          "n_vars": {
            "title": "N Vars",
            "type": "integer"
          },
          "objective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "penalty_components": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Penalty Components",
            "type": "object"
          },
          "self_check": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SolverSelfCheck"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "wall_time_s": {
            "title": "Wall Time S",
            "type": "number"
          }
        },
        "required": [
          "status",
          "wall_time_s",
          "n_vars",
          "n_constraints"
        ],
        "title": "SolveResultPayload",
        "type": "object"
      },
      "SolverSelfCheck": {
        "description": "Automated post-solve hard-constraint audit. Every solve emits one;\na green result is the feasibility receipt researchers and coordinators\ncan trust. A non-green result means the model and the validator\ndisagree \u2014 always a bug worth investigating.",
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "rules_failed": {
            "items": {
              "type": "string"
            },
            "title": "Rules Failed",
            "type": "array"
          },
          "rules_passed": {
            "items": {
              "type": "string"
            },
            "title": "Rules Passed",
            "type": "array"
          },
          "violation_count": {
            "title": "Violation Count",
            "type": "integer"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/SelfCheckViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "violation_count"
        ],
        "title": "SolverSelfCheck",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Healthcare Roster Scheduler v2",
    "version": "2.0.0-phase1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/lab/run": {
      "post": {
        "description": "Execute a cross-product of (solver \u00d7 seed) on the current session\nstate. Blocks until every cell finishes. Callers should keep batches\nmodest (\u2264 3 solvers \u00d7 \u2264 5 seeds \u00d7 \u2264 30 s budget) until Phase 3 adds\na background-job path.",
        "operationId": "run_api_lab_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run",
        "tags": [
          "lab"
        ]
      }
    },
...
    "/api/lab/runs/{batch_id}": {
      "get": {
        "operationId": "get_batch_detail_api_lab_runs__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Batch Detail Api Lab Runs  Batch Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batch Detail",
        "tags": [
          "lab"
        ]
      }
    },
    "/api/lab/runs/{batch_id}/details/{run_id}": {
      "get": {
        "operationId": "get_run_detail_api_lab_runs__batch_id__details__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Detail",
        "tags": [
          "lab"
        ]
      }
    },
...
  }
}